refactor(patient-portal): extract chat session orchestration
