+++ apps/patient-portal/src/app/(app)/chat/page.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import { useEffect, useRef, useState } from "react";
-import { useRouter, useSearchParams } from "next/navigation";
 import {
     HiArrowUp,
     HiChevronDown,
@@ -10,100 +9,36 @@ import {
     HiSparkles,
     HiStop,
 } from "react-icons/hi2";
-import { useDispatch, useSelector } from "react-redux";
 import { ChatBubble } from "@/components/features";
 import { Button, Input } from "@/components/ui";
-import {
-    createSpeechRecognitionController,
-    getVoiceCapabilities,
-    playAssistantVoiceResponse,
-    stopAssistantVoicePlayback,
-    type SpeechRecognitionController,
-    type VoiceStatus,
-} from "@/services/browser-voice";
-import {
-    consumePendingChatDocumentContext,
-    type PendingChatDocumentContext,
-} from "@/services/chat-bridge";
-import {
-    buildChatWebSocketUrl,
-    fetchChatHistory,
-    isChatSocketEvent,
-    mapChatMessageFromApi,
-} from "@/services/chat-api";
-import {
-    addMessage,
-    clearChatState,
-    setChatError,
-    setConnectionStatus,
-    setLoading,
-    setMessages,
-    setTyping,
-} from "@/store/slices/chat-slice";
-import type { AppDispatch, RootState } from "@/store/store";
-import { ChatRole, Language, type ChatMessage, type Language as ChatLanguage } from "@/types";
-
-const CHAT_LANGUAGE_STORAGE_KEY = "patient-portal.chat.language";
+import { usePatientChatSession } from "@/hooks/use-patient-chat-session";
+import { ChatRole, Language, type Language as ChatLanguage } from "@/types";
 
 function getLanguageLabel(language: ChatLanguage): string {
-    return language === Language.ES ? "Español" : "English";
+    return language === Language.ES ? "Espanol" : "English";
 }
 
-function getConnectionLabel(status: RootState["chat"]["connectionStatus"]): string {
-    if (status === "connected") {
+function getConnectionLabel(
+    connectionStatus: "idle" | "connected" | "connecting" | "disconnected" | "error",
+): string {
+    if (connectionStatus === "connected") {
         return "Online";
     }
-    if (status === "connecting") {
+    if (connectionStatus === "connecting") {
         return "Connecting";
     }
-    if (status === "error") {
+    if (connectionStatus === "error") {
         return "Connection issue";
     }
     return "Offline";
-}
-
-function resolveInitialLanguage(): ChatLanguage {
-    if (typeof window === "undefined") {
-        return Language.EN;
-    }
-
-    const stored = window.localStorage.getItem(CHAT_LANGUAGE_STORAGE_KEY);
-    if (stored === Language.ES || stored === Language.EN) {
-        return stored;
-    }
-
-    return window.navigator.language.toLowerCase().startsWith("es")
-        ? Language.ES
-        : Language.EN;
-}
-
-function buildWelcomeMessage(patientId: string, language: ChatLanguage): ChatMessage {
-    return {
-        content:
-            language === Language.ES
-                ? "Hola. Puedo ayudarte a entender resultados, seguir síntomas y preparar preguntas para tu médico."
-                : "Hi. I can help explain results, track symptoms, and prepare questions for your doctor.",
-        createdAt: new Date().toISOString(),
-        id: "welcome-message",
-        language,
-        patientId,
-        role: ChatRole.ASSISTANT,
-    };
-}
-
-function formatSessionTimeLabel(): string {
-    return `Today, ${new Date().toLocaleTimeString("en-US", {
-        hour: "numeric",
-        minute: "2-digit",
-    })}`;
 }
 
 function buildQuickPrompts(language: ChatLanguage): string[] {
     if (language === Language.ES) {
         return [
             "Explica mis resultados recientes",
-            "¿Debo preocuparme por este síntoma?",
-            "Ayúdame a preparar preguntas para mi médico",
+            "Debo preocuparme por este sintoma?",
+            "Ayudame a preparar preguntas para mi medico",
         ];
     }
 
@@ -114,401 +49,50 @@ function buildQuickPrompts(language: ChatLanguage): string[] {
     ];
 }
 
+function formatSessionTimeLabel(): string {
+    return `Today, ${new Date().toLocaleTimeString("en-US", {
+        hour: "numeric",
+        minute: "2-digit",
+    })}`;
+}
+
 export default function ChatPage() {
-    const dispatch = useDispatch<AppDispatch>();
-    const router = useRouter();
-    const searchParams = useSearchParams();
-    const { accessToken, user } = useSelector((state: RootState) => state.auth);
-    const { connectionStatus, error, isTyping, loading, messages } = useSelector(
-        (state: RootState) => state.chat,
-    );
-
-    const [initialDocumentContext] = useState<PendingChatDocumentContext | null>(() => {
-        if (typeof window === "undefined") {
-            return null;
-        }
-
-        const documentId = new URLSearchParams(window.location.search).get("document");
-        return consumePendingChatDocumentContext(documentId);
-    });
-    const [initialLanguage] = useState<ChatLanguage>(
-        () => initialDocumentContext?.preferredLanguage ?? resolveInitialLanguage(),
-    );
-    const voiceCapabilities = getVoiceCapabilities();
-    const selectedLanguageRef = useRef<ChatLanguage>(initialLanguage);
-    const voiceModeRef = useRef(false);
-    const recognitionRef = useRef<SpeechRecognitionController | null>(null);
-    const playbackStopRef = useRef<(() => void) | null>(null);
-    const socketRef = useRef<WebSocket | null>(null);
     const bottomRef = useRef<HTMLDivElement | null>(null);
-
-    const [input, setInput] = useState(
-        () => initialDocumentContext?.suggestedQuestion ?? "",
-    );
     const [sessionTimeLabel] = useState(() => formatSessionTimeLabel());
-    const [selectedLanguage, setSelectedLanguage] =
-        useState<ChatLanguage>(initialLanguage);
-    const [assistantDraft, setAssistantDraft] = useState("");
-    const [assistantDraftStartedAt, setAssistantDraftStartedAt] = useState<string | null>(null);
-    const [voiceModeEnabled, setVoiceModeEnabled] = useState(false);
-    const [voiceStatus, setVoiceStatus] = useState<VoiceStatus>(
-        voiceCapabilities.recognition || voiceCapabilities.synthesis
-            ? "idle"
-            : "unsupported",
-    );
-    const [voiceError, setVoiceError] = useState<string | null>(null);
-    const [voiceInterimTranscript, setVoiceInterimTranscript] = useState("");
-    const [documentContext, setDocumentContext] =
-        useState<PendingChatDocumentContext | null>(initialDocumentContext);
+    const {
+        assistantDraft,
+        assistantDraftStartedAt,
+        canPlayAssistantAudio,
+        connectionStatus,
+        dismissDocumentContext,
+        documentContext,
+        error,
+        handleLanguageSelection,
+        handleMicClick,
+        handlePlayAssistantMessage,
+        handleSend,
+        handleVoiceModeToggle,
+        input,
+        isTyping,
+        loading,
+        messages,
+        selectedLanguage,
+        setInput,
+        voiceError,
+        voiceInterimTranscript,
+        voiceModeEnabled,
+        voiceStatus,
+    } = usePatientChatSession();
 
-    useEffect(() => {
-        bottomRef.current?.scrollIntoView({ behavior: "smooth" });
-    }, [assistantDraft, isTyping, messages]);
-
-    useEffect(() => {
-        selectedLanguageRef.current = selectedLanguage;
-        if (typeof window !== "undefined") {
-            window.localStorage.setItem(CHAT_LANGUAGE_STORAGE_KEY, selectedLanguage);
-        }
-    }, [selectedLanguage]);
-
-    useEffect(() => {
-        voiceModeRef.current = voiceModeEnabled;
-    }, [voiceModeEnabled]);
-
-    useEffect(() => {
-        const documentId = searchParams.get("document");
-        if (!documentId) {
-            return;
-        }
-
-        router.replace("/chat");
-    }, [router, searchParams]);
-
-    useEffect(() => {
-        if (!accessToken || !user) {
-            return;
-        }
-
-        let isMounted = true;
-        dispatch(clearChatState());
-        dispatch(setLoading(true));
-
-        fetchChatHistory(user.id, accessToken)
-            .then((history) => {
-                if (!isMounted) {
-                    return;
-                }
-
-                dispatch(
-                    setMessages(
-                        history.length > 0
-                            ? history
-                            : [buildWelcomeMessage(user.id, initialLanguage)],
-                    ),
-                );
-            })
-            .catch(() => {
-                if (!isMounted) {
-                    return;
-                }
-
-                dispatch(
-                    setMessages([
-                        buildWelcomeMessage(user.id, initialLanguage),
-                    ]),
-                );
-                dispatch(setChatError("Unable to load chat history. Live chat is still available."));
-            })
-            .finally(() => {
-                if (isMounted) {
-                    dispatch(setLoading(false));
-                }
-            });
-
-        dispatch(setConnectionStatus("connecting"));
-        dispatch(setChatError(null));
-
-        const socket = new WebSocket(buildChatWebSocketUrl(user.id, accessToken));
-        socketRef.current = socket;
-
-        socket.onopen = () => {
-            if (!isMounted) {
-                return;
-            }
-            dispatch(setConnectionStatus("connected"));
-            dispatch(setChatError(null));
-        };
-
-        socket.onmessage = (event) => {
-            let payload: unknown;
-
-            try {
-                payload = JSON.parse(event.data as string);
-            } catch {
-                return;
-            }
-
-            if (!isChatSocketEvent(payload)) {
-                return;
-            }
-
-            switch (payload.type) {
-                case "chat_history": {
-                    const incomingHistory = payload.messages.map(mapChatMessageFromApi);
-                    dispatch(
-                        setMessages(
-                            incomingHistory.length > 0
-                                ? incomingHistory
-                                : [buildWelcomeMessage(user.id, selectedLanguageRef.current)],
-                        ),
-                    );
-                    return;
-                }
-                case "user_message_saved":
-                    dispatch(addMessage(mapChatMessageFromApi(payload.message)));
-                    return;
-                case "assistant_start":
-                    setAssistantDraft("");
-                    setAssistantDraftStartedAt(new Date().toISOString());
-                    dispatch(setTyping(true));
-                    return;
-                case "assistant_chunk":
-                    setAssistantDraft((current) => `${current}${payload.content}`);
-                    dispatch(setTyping(true));
-                    return;
-                case "assistant_complete": {
-                    const assistantMessage = mapChatMessageFromApi(payload.message);
-                    setAssistantDraft("");
-                    setAssistantDraftStartedAt(null);
-                    dispatch(setTyping(false));
-                    dispatch(addMessage(assistantMessage));
-
-                    if (voiceModeRef.current) {
-                        playbackStopRef.current?.();
-                        playbackStopRef.current = playAssistantVoiceResponse({
-                            audioUrl: assistantMessage.audioUrl,
-                            language: assistantMessage.language,
-                            onEnd: () => {
-                                setVoiceStatus(
-                                    voiceModeRef.current ? "idle" : "unsupported",
-                                );
-                            },
-                            onStart: () => {
-                                setVoiceStatus("playing");
-                            },
-                            text: assistantMessage.content,
-                        });
-                    }
-
-                    if (payload.escalation_required) {
-                        dispatch(
-                            setChatError(
-                                "Urgent symptoms detected. Contact your care team today.",
-                            ),
-                        );
-                    }
-                    return;
-                }
-                case "escalation_recommended":
-                    dispatch(setChatError(payload.message));
-                    return;
-                case "error":
-                    setAssistantDraft("");
-                    setAssistantDraftStartedAt(null);
-                    dispatch(setTyping(false));
-                    dispatch(setChatError(payload.message || "Chat request failed."));
-                    return;
-                default:
-                    return;
-            }
-        };
-
-        socket.onerror = () => {
-            if (!isMounted) {
-                return;
-            }
-            setAssistantDraft("");
-            setAssistantDraftStartedAt(null);
-            dispatch(setConnectionStatus("error"));
-            dispatch(setTyping(false));
-            dispatch(setChatError("Live chat connection encountered an issue."));
-        };
-
-        socket.onclose = () => {
-            if (!isMounted) {
-                return;
-            }
-            dispatch(setConnectionStatus("disconnected"));
-            dispatch(setTyping(false));
-        };
-
-        return () => {
-            isMounted = false;
-            recognitionRef.current?.stop();
-            playbackStopRef.current?.();
-            playbackStopRef.current = null;
-            socketRef.current = null;
-            socket.close();
-        };
-    }, [accessToken, dispatch, initialLanguage, user]);
-
-    useEffect(() => {
-        return () => {
-            recognitionRef.current?.stop();
-            stopAssistantVoicePlayback();
-        };
-    }, []);
-
-    function dismissDocumentContext() {
-        setDocumentContext(null);
-    }
-
-    function sendChatMessage(content: string, audioUrl?: string): void {
-        const trimmedContent = content.trim();
-        if (!trimmedContent) {
-            return;
-        }
-
-        if (!socketRef.current || socketRef.current.readyState !== WebSocket.OPEN) {
-            dispatch(setChatError("Chat is reconnecting. Try sending in a moment."));
-            return;
-        }
-
-        socketRef.current.send(
-            JSON.stringify({
-                type: "user_message",
-                content: trimmedContent,
-                language: selectedLanguageRef.current,
-                audio_url: audioUrl ?? null,
-            }),
-        );
-
-        dispatch(setChatError(null));
-        setVoiceError(null);
-        setVoiceInterimTranscript("");
-        setInput("");
-    }
-
-    function handleSend(event: React.FormEvent<HTMLFormElement>) {
-        event.preventDefault();
-        sendChatMessage(input);
-    }
-
-    function handleLanguageSelection(language: ChatLanguage) {
-        setSelectedLanguage(language);
-        setVoiceError(null);
-    }
-
-    function handleVoiceModeToggle() {
-        if (voiceStatus === "unsupported") {
-            setVoiceError("Voice mode is not available on this device.");
-            return;
-        }
-
-        const nextValue = !voiceModeEnabled;
-        setVoiceModeEnabled(nextValue);
-
-        if (!nextValue) {
-            recognitionRef.current?.stop();
-            playbackStopRef.current?.();
-            playbackStopRef.current = null;
-            setVoiceInterimTranscript("");
-            setVoiceStatus("idle");
-        } else {
-            setVoiceError(null);
-        }
-    }
-
-    function handleMicClick() {
-        if (!voiceCapabilities.recognition) {
-            setVoiceStatus("unsupported");
-            setVoiceError("Speech recognition is not supported in this browser.");
-            return;
-        }
-
-        if (voiceStatus === "listening") {
-            recognitionRef.current?.stop();
-            return;
-        }
-
-        setVoiceError(null);
-        setVoiceInterimTranscript("");
-
-        const controller = createSpeechRecognitionController(selectedLanguageRef.current, {
-            onEnd: (finalTranscript) => {
-                recognitionRef.current = null;
-                setVoiceStatus(voiceModeRef.current ? "processing" : "idle");
-                setVoiceInterimTranscript("");
-
-                if (!finalTranscript) {
-                    setVoiceStatus("idle");
-                    return;
-                }
-
-                if (voiceModeRef.current) {
-                    sendChatMessage(finalTranscript);
-                } else {
-                    setInput(finalTranscript);
-                    setVoiceStatus("idle");
-                }
-            },
-            onError: (message) => {
-                recognitionRef.current = null;
-                setVoiceStatus("idle");
-                setVoiceInterimTranscript("");
-                setVoiceError(message);
-            },
-            onStart: () => {
-                setVoiceStatus("listening");
-            },
-            onTranscript: ({ finalTranscript, interimTranscript }) => {
-                if (!voiceModeRef.current && finalTranscript) {
-                    setInput(finalTranscript);
-                }
-                setVoiceInterimTranscript(interimTranscript);
-            },
-        });
-
-        if (!controller) {
-            setVoiceStatus("unsupported");
-            setVoiceError("Speech recognition is not supported in this browser.");
-            return;
-        }
-
-        recognitionRef.current = controller;
-        controller.start();
-    }
-
-    function handlePlayAssistantMessage(message: ChatMessage) {
-        const canPlay =
-            Boolean(message.audioUrl) || voiceCapabilities.synthesis;
-
-        if (!canPlay) {
-            setVoiceError("Audio playback is not available on this device.");
-            return;
-        }
-
-        playbackStopRef.current?.();
-        playbackStopRef.current = playAssistantVoiceResponse({
-            audioUrl: message.audioUrl,
-            language: message.language,
-            onEnd: () => {
-                setVoiceStatus("idle");
-            },
-            onStart: () => {
-                setVoiceStatus("playing");
-            },
-            text: message.content,
-        });
-    }
-
-    const canPlayAssistantAudio = voiceCapabilities.synthesis;
     const quickPrompts = buildQuickPrompts(selectedLanguage);
     const showQuickPrompts =
         !loading
         && !documentContext
-        && messages.filter((message) => message.role === ChatRole.USER).length === 0;
+        && !messages.some((message) => message.role === ChatRole.USER);
+
+    useEffect(() => {
+        bottomRef.current?.scrollIntoView({ behavior: "smooth" });
+    }, [assistantDraft, isTyping, messages]);
 
     return (
         <div className="min-h-full bg-[#F5F8FE] px-3 py-4 text-[#23324A] sm:px-6 sm:py-6">
@@ -623,7 +207,7 @@ export default function ChatPage() {
                             <div className="space-y-3">
                                 <div className="rounded-[22px] border border-[#E3EBF7] bg-white px-4 py-3 text-sm text-[#41536F] shadow-[0_16px_32px_rgba(70,96,140,0.08)]">
                                     {selectedLanguage === Language.ES
-                                        ? "Puedo ayudarte con síntomas, resultados y próximos pasos. Prueba una de estas preguntas:"
+                                        ? "Puedo ayudarte con sintomas, resultados y proximos pasos. Prueba una de estas preguntas:"
                                         : "I can help with symptoms, results, and next steps. Try one of these prompts:"}
                                 </div>
                                 <div className="flex flex-col gap-2">
