+++ apps/patient-portal/src/app/(app)/records/page.tsx
@@ -15,7 +15,11 @@ import {
 } from "@/services/chat-bridge";
 import { uploadDocumentToStorage } from "@/services/storage";
 import type { RootState } from "@/store/store";
-import { DocumentType, Language, type Document } from "@/types";
+import {
+    DocumentType,
+    Language,
+    type Document,
+} from "@/types";
 import { useSelector } from "react-redux";
 
 type PortalDocument = Document & { icon: ReactNode; provider: string };
@@ -24,7 +28,7 @@ interface DocumentApiRecord {
     id: string;
     patient_id: string;
     uploaded_by: string;
-    uploaded_by_role: "patient" | "clinician";
+    uploaded_by_role: Document["uploadedByRole"];
     document_type: DocumentType;
     file_name: string;
     file_url: string;
@@ -36,7 +40,7 @@ interface DocumentApiRecord {
     parse_error?: string | null;
     parse_attempts?: number;
     source_clinic?: string | null;
-    visibility: "all_providers" | "specific_provider";
+    visibility: Document["visibility"];
     created_at: string;
 }
 
@@ -115,7 +119,7 @@ export default function RecordsPage() {
     const router = useRouter();
     const [documents, setDocuments] = useState<PortalDocument[]>([]);
     const [selectedDocument, setSelectedDocument] = useState<PortalDocument | null>(null);
-    const [explanationLang, setExplanationLang] = useState<"en" | "es">("en");
+    const [explanationLang, setExplanationLang] = useState<Language>(Language.EN);
     const [explanationText, setExplanationText] = useState<string | null>(null);
     const [explanationLoading, setExplanationLoading] = useState(false);
     const [loading, setLoading] = useState(true);
@@ -256,8 +260,13 @@ export default function RecordsPage() {
         }
     }
 
-    async function handleLanguageChange(lang: "en" | "es") {
-        setExplanationLang(lang);
+    async function handleLanguageChange(language: Language) {
+        setExplanationLang(language);
+        if (language === Language.EN && selectedDocument?.parseStatus === "completed") {
+            setExplanationText(selectedDocument.aiSummary ?? null);
+            return;
+        }
+
         if (!selectedDocument) {
             return;
         }
@@ -266,16 +275,11 @@ export default function RecordsPage() {
             return;
         }
 
-        if (lang === "en") {
-            setExplanationText(selectedDocument.aiSummary ?? null);
-            return;
-        }
-
         setExplanationLoading(true);
         try {
             const result = await api.post<{ summary: string }>(
                 `/api/v1/documents/${selectedDocument.id}/explain`,
-                { language: lang },
+                { language },
                 { token: accessToken ?? undefined },
             );
             setExplanationText(result.summary);
@@ -291,19 +295,16 @@ export default function RecordsPage() {
             return;
         }
 
-        const preferredLanguage =
-            explanationLang === "es" ? Language.ES : Language.EN;
-
         storePendingChatDocumentContext({
             documentId: selectedDocument.id,
             documentName: selectedDocument.fileName,
             documentType: selectedDocument.documentType,
-            preferredLanguage,
+            preferredLanguage: explanationLang,
             provider: selectedDocument.provider,
             suggestedQuestion: buildSuggestedDocumentQuestion({
                 documentName: selectedDocument.fileName,
                 documentType: selectedDocument.documentType,
-                preferredLanguage,
+                preferredLanguage: explanationLang,
                 provider: selectedDocument.provider,
             }),
             summary: explanationText ?? selectedDocument.aiSummary,
@@ -399,7 +400,7 @@ export default function RecordsPage() {
                         name={document.fileName}
                         onClick={() => {
                             setSelectedDocument(document);
-                            setExplanationLang("en");
+                            setExplanationLang(Language.EN);
                             setExplanationLoading(false);
                             if (document.parseStatus === "failed") {
                                 setExplanationText(
@@ -441,11 +442,11 @@ export default function RecordsPage() {
                             <p className="text-xs font-semibold uppercase tracking-wide text-blue-600">Explain this to me</p>
                             <select
                                 className="rounded-lg border border-blue-200 bg-white px-2 py-1 text-xs text-blue-700"
-                                onChange={(event) => handleLanguageChange(event.target.value as "en" | "es")}
+                                onChange={(event) => handleLanguageChange(event.target.value as Language)}
                                 value={explanationLang}
                             >
-                                <option value="en">English</option>
-                                <option value="es">Español</option>
+                                <option value={Language.EN}>English</option>
+                                <option value={Language.ES}>Español</option>
                             </select>
                         </div>
                         <p className="mt-2 text-sm text-gray-700">
